Use `CardinalPointGenerator` in place of `PointGenerator1D`
`PointGenerator1D` is meant as a orientation-independent
declarative way to create points in relation to the
bounding boxes. You give an orientation (vertical/horizontal)
to the constructor the points generated will be appropriatly
mirrored. This is super useful in drawing axes since the same
code can be reused once you've configured the point generator
with the right orientation. However, the way it creates point,
while succint, is confusing to reason about.

`CardinalPointGenerator` on the other is much more simpler since
it refers to the bounding box's border in cardinal directions.
The drawback is that you're gonna have to write different code for
different orientations. But I think this is a net win since the
resulting code is a lot easier to understand.