Delegate SVG's flipped y-axis issue to drawing classes
SVG viewBox's origin is at the topleft corner, which causes
issues when drawing a chart, whose origin usually resides at
the bottomleft corner. Thus, it is necessary to flip the y-axis
of the dataset somewhere in the rendering pipeline.

Previously this flipping is scattered throughout the codebase at
random points.

Now, the necessary flipping is solely done by the classes that
actually draw the chart (e.g. `LinePen`, `TextPen` etc.). This
enables us to not worry about the flipped y-axis when we're
doing business logic.